Captures all network exceptions in a new module. Adds a BadMessageException so that it can be better traced.
